fix: Align Popup light-dismiss with WinUI
Handle pointer events exclusively on PopupPanel (not quite PopupRoot, but close). Rather than bounds-checking and adding handlers to the child content, just check the OriginalSource of the event, as WinUI does, to exclude events that are 'blocked' by the Popup content. This aligns Uno with WinUI in terms of hit-transparent areas of the content.